requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.2

